bookmarks: move storage key names to utils.js
Because these names would also be needed in options section.
